+++ assemblyadapter-common-recycler-divider/src/main/java/com/github/panpf/assemblyadapter/recycler/divider/internal/StaggeredGridDividerItemDecorationHelper.kt
@@ -19,8 +19,6 @@ import android.graphics.Canvas
 import android.graphics.Rect
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
-import kotlin.math.ceil
-import kotlin.math.floor
 
 class StaggeredGridDividerItemDecorationHelper(private val itemDividerProvider: StaggeredGridItemDividerProvider) {
 
@@ -40,100 +38,105 @@ class StaggeredGridDividerItemDecorationHelper(private val itemDividerProvider: 
     ) {
         val isFirstSpan = isFullSpan || spanIndex == 0
         val isLastSpan = isFullSpan || spanIndex == spanCount - 1
+        val spanSize = if (isFullSpan) spanCount else 1
 
         val startItemDivider = itemDividerProvider.getItemDivider(
             view, parent, itemCount, position, spanCount, spanIndex, isFullSpan, isFirstSpan,
             isLastSpan, isColumnFirst, isColumnEnd, isVerticalOrientation,
-            if (isLTRDirection) ItemDivider.Type.START else ItemDivider.Type.END
+            if (isLTRDirection) ItemDivider.Type.START else ItemDivider.Type.END, false
         )
         val topItemDivider = itemDividerProvider.getItemDivider(
             view, parent, itemCount, position, spanCount, spanIndex, isFullSpan, isFirstSpan,
-            isLastSpan, isColumnFirst, isColumnEnd, isVerticalOrientation, ItemDivider.Type.TOP
+            isLastSpan, isColumnFirst, isColumnEnd, isVerticalOrientation,
+            ItemDivider.Type.TOP, false
         )
         val endItemDivider = itemDividerProvider.getItemDivider(
             view, parent, itemCount, position, spanCount, spanIndex, isFullSpan, isFirstSpan,
             isLastSpan, isColumnFirst, isColumnEnd, isVerticalOrientation,
-            if (isLTRDirection) ItemDivider.Type.END else ItemDivider.Type.START
+            if (isLTRDirection) ItemDivider.Type.END else ItemDivider.Type.START, false
         )
         val bottomItemDivider = itemDividerProvider.getItemDivider(
             view, parent, itemCount, position, spanCount, spanIndex, isFullSpan, isFirstSpan,
-            isLastSpan, isColumnFirst, isColumnEnd, isVerticalOrientation, ItemDivider.Type.BOTTOM
+            isLastSpan, isColumnFirst, isColumnEnd, isVerticalOrientation,
+            ItemDivider.Type.BOTTOM, false
         )
-        val startItemDividerSize = startItemDivider?.widthSize ?: 0
-        val topItemDividerSize = topItemDivider?.heightSize ?: 0
-        val endItemDividerSize = endItemDivider?.widthSize ?: 0
-        val bottomItemDividerSize = bottomItemDivider?.heightSize ?: 0
 
+        // 当我们希望显示 sideDivider, sideHeaderAndFooterDivider 并且 item 的宽度是 parent 的宽度减去所有 divider 后除以 spanCount 时，
+        // 如公式：'val itemSize=(parentWidth - (dividerSize * (spanCount+1))) / spanCount'
+        // 按照 GridItemDividerProvider 的逻辑，第一个 item 的 start 和 end 将都会有 divider 显示
+        // 因为 GridLayoutManager 强制每个 item 的最大宽度为 parentWidth/spanCount，
+        // 所以第一个 item 的宽度会因为加上 start 和 end 的 divider 后超过 GridLayoutManager 限制的最大宽度，
+        // 这时 GridLayoutManager 会将 item 的宽度修改为最大宽度减去 start 和 end 的 divider，导致 item 最终的宽度不是我们希望的宽度
+        // 所以以下的代码都是为了解决这个问题
+        val showHeaderAndFooterSideDivider =
+            itemDividerProvider.sideHeaderAndFooterDividerConfig != null
         if (isVerticalOrientation) {
-            when {
-                isFullSpan -> {
-                    outRect.set(
-                        startItemDividerSize,
-                        topItemDividerSize,
-                        endItemDividerSize,
-                        bottomItemDividerSize
+            val sideDividerSize = startItemDivider?.widthSize ?: endItemDivider?.widthSize
+            val startItemDividerSize =
+                when {
+                    sideDividerSize == null -> 0
+                    isFirstSpan -> sideDividerSize
+                    else -> normalizedOffsetFromSize(
+                        ItemDivider.Type.START,
+                        sideDividerSize,
+                        spanCount,
+                        spanIndex + spanSize - 1,
+                        showHeaderAndFooterSideDivider
                     )
                 }
-                isFirstSpan -> {
-                    outRect.set(
-                        startItemDividerSize,
-                        topItemDividerSize,
-                        floor(endItemDividerSize / 2f).toInt(),
-                        bottomItemDividerSize
-                    )
-                }
-                isLastSpan -> {
-                    outRect.set(
-                        ceil(startItemDividerSize / 2f).toInt(),
-                        topItemDividerSize,
-                        endItemDividerSize,
-                        bottomItemDividerSize
-                    )
-                }
-                else -> {
-                    outRect.set(
-                        ceil(startItemDividerSize / 2f).toInt(),
-                        topItemDividerSize,
-                        floor(endItemDividerSize / 2f).toInt(),
-                        bottomItemDividerSize
-                    )
-                }
+            val endItemDividerSize = when {
+                sideDividerSize == null -> 0
+                isLastSpan -> sideDividerSize
+                else -> normalizedOffsetFromSize(
+                    ItemDivider.Type.END,
+                    sideDividerSize,
+                    spanCount,
+                    spanIndex + spanSize - 1,
+                    showHeaderAndFooterSideDivider
+                )
             }
+            val topItemDividerSize = topItemDivider?.heightSize ?: 0
+            val bottomItemDividerSize = bottomItemDivider?.heightSize ?: 0
+
+            outRect.set(
+                startItemDividerSize,
+                topItemDividerSize,
+                endItemDividerSize,
+                bottomItemDividerSize
+            )
         } else {
-            when {
-                isFullSpan -> {
-                    outRect.set(
-                        startItemDividerSize,
-                        topItemDividerSize,
-                        endItemDividerSize,
-                        bottomItemDividerSize
-                    )
-                }
-                isFirstSpan -> {
-                    outRect.set(
-                        startItemDividerSize,
-                        topItemDividerSize,
-                        endItemDividerSize,
-                        floor(bottomItemDividerSize / 2f).toInt()
-                    )
-                }
-                isLastSpan -> {
-                    outRect.set(
-                        startItemDividerSize,
-                        ceil(topItemDividerSize / 2f).toInt(),
-                        endItemDividerSize,
-                        bottomItemDividerSize
-                    )
-                }
-                else -> {
-                    outRect.set(
-                        startItemDividerSize,
-                        ceil(topItemDividerSize / 2f).toInt(),
-                        endItemDividerSize,
-                        floor(bottomItemDividerSize / 2f).toInt()
-                    )
-                }
+            val sideDividerSize = topItemDivider?.heightSize ?: bottomItemDivider?.heightSize
+            val topItemDividerSize = when {
+                sideDividerSize == null -> 0
+                isFirstSpan -> sideDividerSize
+                else -> normalizedOffsetFromSize(
+                    ItemDivider.Type.TOP,
+                    sideDividerSize,
+                    spanCount,
+                    spanIndex + spanSize - 1,
+                    showHeaderAndFooterSideDivider
+                )
             }
+            val bottomItemDividerSize = when {
+                sideDividerSize == null -> 0
+                isLastSpan -> sideDividerSize
+                else -> normalizedOffsetFromSize(
+                    ItemDivider.Type.BOTTOM,
+                    sideDividerSize,
+                    spanCount,
+                    spanIndex + spanSize - 1,
+                    showHeaderAndFooterSideDivider
+                )
+            }
+            val startItemDividerSize = startItemDivider?.widthSize ?: 0
+            val endItemDividerSize = endItemDivider?.widthSize ?: 0
+
+            outRect.set(
+                startItemDividerSize,
+                topItemDividerSize,
+                endItemDividerSize,
+                bottomItemDividerSize
+            )
         }
     }
 
@@ -157,31 +160,22 @@ class StaggeredGridDividerItemDecorationHelper(private val itemDividerProvider: 
         val startItemDivider = itemDividerProvider.getItemDivider(
             view, parent, itemCount, position, spanCount, spanIndex, isFullSpan, isFirstSpan,
             isLastSpan, isColumnFirst, isColumnEnd, isVerticalOrientation,
-            if (isLTRDirection) ItemDivider.Type.START else ItemDivider.Type.END
+            if (isLTRDirection) ItemDivider.Type.START else ItemDivider.Type.END, true
         )
         val topItemDivider = itemDividerProvider.getItemDivider(
             view, parent, itemCount, position, spanCount, spanIndex, isFullSpan, isFirstSpan,
-            isLastSpan, isColumnFirst, isColumnEnd, isVerticalOrientation, ItemDivider.Type.TOP
+            isLastSpan, isColumnFirst, isColumnEnd, isVerticalOrientation,
+            ItemDivider.Type.TOP, true
         )
         val endItemDivider = itemDividerProvider.getItemDivider(
             view, parent, itemCount, position, spanCount, spanIndex, isFullSpan, isFirstSpan,
             isLastSpan, isColumnFirst, isColumnEnd, isVerticalOrientation,
-            if (isLTRDirection) ItemDivider.Type.END else ItemDivider.Type.START
+            if (isLTRDirection) ItemDivider.Type.END else ItemDivider.Type.START, true
         )
         val bottomItemDivider = itemDividerProvider.getItemDivider(
-            view,
-            parent,
-            itemCount,
-            position,
-            spanCount,
-            spanIndex,
-            isFullSpan,
-            isFirstSpan,
-            isLastSpan,
-            isColumnFirst,
-            isColumnEnd,
-            isVerticalOrientation,
-            ItemDivider.Type.BOTTOM
+            view, parent, itemCount, position, spanCount, spanIndex, isFullSpan, isFirstSpan,
+            isLastSpan, isColumnFirst, isColumnEnd, isVerticalOrientation,
+            ItemDivider.Type.BOTTOM, true
         )
         val startItemDividerSize = startItemDivider?.widthSize ?: 0
         val topItemDividerSize = topItemDivider?.heightSize ?: 0